requirements.txt
mariadb==1.1.6
packaging==23.1
python-dotenv==1.0.0
